Missed some changes to preprocessor definitions, include directories and linker inputs  - (followup to 1405a19700af47b25d9bfccb45db1a6914d651a6)
A note on iterator debugging:
My Boost 1.55 binaries were built with it enabled, I had no control over that, so I had to enable it for the debug builds, otherwise it wouldn't build because of a mismatch.
From what I gathered it can impact the performance of the debugger because all the extra stuff it needs to do but I can't confirm.
I tried the "_ALLOW_ITERATOR_DEBUG_LEVEL_MISMATCH" definition in hopes to preserve the current behavior, but this ended up with the exes insta-crashing.
This is life now.