Removed TransfersDamageToParent from Attachable, just using MOSRotating::DamageMultiplier now. Added flag for whether this was set in INI so hardcoded Attachables can properly modify it as needed.
Added Attachable InheritsHFlipped which can be 0 (no inheritance), 1 (inherit directly) or any other number, though 2 is recommended (inherit reversed). Could be split into 2 variables, but I figured this was cleaner, if a little weird. Didn't wanna use -1 since we often use that to mean default value.
Changed wound handling stuff to allow for searching for wounds that have a positive damage multipliers, a negative damage multiplier or no damage multiplier, so it can be more fine-grained