Changed hardcoded attachable handling to put lambda functions in a map on MOSR so that they can later be used to unset hardcoded attachable pointers when needed
Changed the already copied attachables unique IDs unordered_set to an unordered_map HardcodedAttachableUniqueIDsAndSetters and used this to store hardcoded attachable setters
Changed the process of MOSR::Create(reference) so once hardcoded attachables are copied over, they sit in the map instead of the reference's versions, making it useable for unsetting hardcoded attachable pointers
Cleaned up a bit of attachable removal so it's more correct about deleting the attachable being removed or adding it to MovableMan, and also so it unsets hardcoded attachable pointers where relevant.
Cleaned up unused extra methods in Turret that RemoveAttachable can handle nicely, made it better align with everything else and also did a bit of cleanup in MOSR
Fixed a mistake in ACDropship with hatches, fixed a mistake in ACRocket with main thruster inheriting rotation, fixed a mistake in Emitter which was screwing up flash
Updated objects with hardcoded attachables accordingly. Also, by necessity, made it so objects read in will add and remove hardcoded attachables to/from their attachable list, with all the handling that comes along with that.