Fixed Attachable terrain collision handling so it properly adds atoms with the proper subgroup ID, and updates them only when appropriate
Fixed travel handling of MOToNotHit in MOPixel, MOSParticle and MOSRotating so it doesn't rely on sequentially ordered IDs and instead uses GetMOIDs so it's safer.
Fixed AEmitter's flash and HDFirearm's flash and magazine and Leg's Foot colliding with terrain while attached
Removed HeldDevice override DeleteWhenRemovedFromParent. It was making Crab Turret HDFirearms not delete, and would generally not be good to force like that. This can have some messy side-effects if a HeldDevice is set to delete when removed though, cause swapping weapons and stuff will delete it. Probably need to add an extra property on HD that allows for selective overriding of this.
Renamed and cleaned up some stuff