Made GetMass more efficient for MOSRs. Instead of having to get each attachable and wound's mass (and in turn their attachables' masses) each call, it stores the mass of each attachable when adding them and removes it when removing them
To support this change to GetMass, attachables are now responsible for telling their parent when their mass has changed because of their own wounds or attachables.
Cleaned up some comments in MOSR and cleaned up Attachable a little
Cleaned up some leftover cruft in TDExplosive
Renamed DetachOrDestroyAll to RemoveOrDestroyAllAttachables and fixed it so it uses an iterator for loop again and won't cause crashes. Would be nice to just delete it since it's only used for fuggin hats