Fixed crashes and potential crashes
Changed RemoveAttachable deletion handling so attachables are always added to MovableMan if they're getting deleted, but can also not be deleted or added to MovableMan so things can utilize them. I didn't realize it, for the previous version to work safely the MOID layer would have to be cleared and redrawn after deleting the attachable in order to avoid screwups. This option is definitely cheaper. Also changed a size() check to an empty() check (thanks sonarlint for catching it)
Added special luaman handling for removing attachables so that modder mistakes can't cause memory leaks (i.e. they will always either delete or get added to MovableMan).
Modified comments and added a TODO to MOSR header to make RemoveAttachable return the attachable removed and hand over ownership. When this is done it will avoid some mess we currently have. It should be done as part of Arm cleanup.