Make Esc Key close open BuyMenus if they exist instead of pausing the game
Two small changes, one to the BuyMenuGUI, closing an open BuyMenu if if the UInput Manager detects any Start press (ie ESC). This will act on all open BuyMenus.
The other is a modification to the UInputMan itself, making sure that if there are any BuyMenus visible, Escape will not pause the Activity. Achieved by getting and casting the GameActivity, and iterating through it's BuyMenuGUI's for each player up to the maximim. All must not be visible for the escape key to pause the activity.