Changed rounding mode of mass tooltips & label to always ceil.
Purchasable mass, when shown as an items tooltip or as the total label now always rounds up the floating point value to it's set precision to ensure that an order of `100.0001/100` which would normally display as `100.0/100.0` now correctly shows as unpurchasable, ie `100.1/100.0`. Same for tooltips, to ensure that an item isn't inaccurately represented to the player as purchasable with current limit, while not being allowed to confirm the order.
Also changed readout of Passenger Label for craft with no passenger space to fit allowed space in label box.