Fixed a longstanding bug data made wrt ordering of drawing when drawing attachables and wounds that aren't drawn after parent. This fixes a different issue with MOID layer drawing for some doors
This was causing another very odd edge-case bug with drawing to MOID layer where non-hardcoded attachables would overwrite their parents' sprites when drawing to non-color layers. Because attachables were infrequently used in the past and the nature of this bug meant it only happened in specific situations, it was probably not very noticeable, but fixing it solves problems with doors and probably other problems.