Changed `HDFirearm: ActiveSound` to pointer.
SoundContainer for HDFirearm's ActiveSound is now a pointer. Added getter/setter methods for this property.
Added checks to ensure that the ActiveSound exists before start/stopping it, including to the externally accessed `StopActivationSound( )`.
Used `StopActivationSound( )` in `HDFirearm::Update( )` as required to stop active sounds, replacing direct calls to the soundcontainer.