Changed `HDFirearm: DeactivationSound` to pointer.
SoundContainer for HDFirearm's ActiveSound is now a pointer. Added getter/setter methods for this property.
Removed typo in reader for PreFireSound erroneously setting DeactivationSound's overlap mode.
Had to rework DeactivationSound as it's barely implemented correctly. Now, only plays when the weapon is inactive and fired last frame. Still needs work, as it's possible to not spawn a round and still have it played. (Seen when heat lance is fired for short period, it produces the prefire and deactivation sounds without actually firing a shot.