Changed `MOSRotating: GibSound` to pointer.
SoundContainer for this property is now stored as a pointer, and is now referenced as such.
Added getter & setter for this property.
Added condition to `ReadProperty( )` that checks if GibSound exists before adding it, as it's overwritten by DetonationSound ini prop of TDExplosives. As such, whichever is defined first is used. This was done to save possible double call to new, and hence possible unfree'd memory allocation.