Made Turret support an unlimited number of mounted HeldDevices
There's still special handling for the first mounted device, to keep things simple for lua and cases that only care about one mounted device. Given this, the first should be considered the "primary" mounted device
Renamed MountedDeviceRotOffset to MountedDeviceRotationOffset cause I felt like it, and added Lua and INI access to it.
Added ACrab handling for this change, some things use the first mounted device, other things apply to all of them. Also cleaned up some minor things in ACrab while I was there
Added another hardcoded attachable map to MOSRotating, this one is for Remover functions that need the attachable to be passed into them, which allows for collections of hardcoded attachables. Used it when removing attachables and destroying MOSR (needed in the latter case cause the Remover won't actually delete the object, it'll just clean it out of the list and relinquish ownership of it, this is necessary to add things to MovableMan safely)