Code review fixes
Fix keyboard handling for inventory menu so you can't go weirdly out of bounds, and made down and right handling more consistent with each other. This probably stops the crash too.
Also tweaked keyboard handling of the menu so it's a little more convenient - moving down when it'd put you past usable buttons will put you at the end, and moving down/right won't let you use empty buttons if it shouldn't
Fix some memory leaks in GameActivity and made it so swapping actors disables the inventory if it's open (also tweaked inventory to not have problems with that)