Unbroke all the mods :P
Added flag for whether or not either throttle was redefined and, if it wasn't, defaulted it to 0 for AHumans' and ACrabs' jetpacks.
Added setter for throttle values, and renamed the getters to be more consistent (though the throttle variables and getters and setters all need to be renamed at some point, there's an issue in for that)
Added lua bindings for min and max throttle range. They've got garbage names because I don't want them bound til we get a better name for this stuff