Fixed bug where pie menu interactions with Actor was ignored (change weapon, drop weapon, etc)
This is a tricky one. It happens because AHuman called Actor::Update() in the middle, after the drop weapon logic (so the command from GameActivity() wasn't overwritten).

I have changed this to instead split the MoveableMan update into phases, so we can insert the controller update into the correct place to ensure the behaviour is similar to before.