ci: refactor meson build workflow to be reusable
Rework the meson build workflow to act as a callable workflow in a composite job.

Disable triggering on pushes to dev/master.
Add input parameters to call this workflow with, that dictate:
- whether to upload build artefacts from this workflow
- which meson build type to use (release/debug)
- which debug level, if any, to compile the code at.

Switched `meson 0.6.0 => 0.6.3` to match the version in our CI script.
Split Meson setup/configuration to accommodate upload/non-upload invocations of the workflow
Configured the upload step to fail with errors if no artefacts are found to be uploaded (when the relevant setting is configured).